seedファイル作成

+++ db/migrate/20230507123644_create_number_of_times.rb
@@ -3,7 +3,7 @@ class CreateNumberOfTimes < ActiveRecord::Migration[6.1]
     create_table :number_of_times do |t|
 
       t.string  :name
-      t.integer :dosing_times
+      t.decimal :dosing_times
 
       t.timestamps
     end
